Add example project
This is the basic starter for spaceflights (based on Kedro 18.4, so it is just two pipelines with no namespaces/modular pipeline stuff in it).

I've added a third pipeline to the starter code which I've called "reporting" and this is the test/example for the plotly/matplotlib charts in kedro viz.